refactor arduino code to be isndie all the same class and abstraction.
